Refactor: 관심사에 맞게 상태 위치 변경
1. App.jsx에서 상태코드 제거
2. 각 관심사에 맞게 핸들러와 데이터 위치 이동
