+++ src/Restaurants.jsx
@@ -1,4 +1,12 @@
-export default function Restaurants({ restaurants, currentRegionId, currentCategoryId }) {
+import { useSelector } from 'react-redux';
+
+export default function Restaurants() {
+  const { restaurants, currentRegionId, currentCategoryId } = useSelector((state) => ({
+    restaurants: state.restaurants,
+    currentRegionId: state.currentRegionId,
+    currentCategoryId: state.currentCategoryId,
+  }));
+
   return (
     <>
       {(currentCategoryId && currentRegionId) && (
